fix: Allow linting of typescript with module resolution set to `node16` or `nodenext`
Node16 module resolution with Typescript requires every relative import to have
a file extension, but Typescript does not allow `.ts` so you have to use `.js`.
As this file obviously does not exist on build time, but only after it has been created
ESLint will complain about an unresolved import. Fixing it by using a different resolver for this.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>