feat(imports): add custom plugin to suggest file extensions
Currently we only forbid file extension less imports / exports.
But when migrating large projects this is cumbersome,
so instead this is a rather small plugin that just checks if relative
imports or exports have file extensions.
If possible it will lookup the file extension from the file system,
so no magic resolver is needed.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>